binding func internally + refactor
